requirement.txt
pandas
psycopg2-binary
python-dotenv
tqdm